wrench: Updated rollup configuration
This updates the rollup config to avoid having the spread operator
in the generated code as that causes issues when people want to
bundle our code with webpack (which doesn't support the spread
operator for object literals).

This also changes the browserbuild script to avoid issues for people
running this on Windows. Hat-tip to @philltew for this one.

Note that this is the first plugin where I'm updating the rollup
config. I'll have to do the same for all plugins and patterns,
but it's unlikely I'll write this long a commit message for each
one of them :)