:wrench: Updated rollup config
This updates the rollup config to avoid using spread operator
in the output code because it causes issues when bundling our
code in projects that use webpack (as webpack does not support
the spread operator in object literals).

It also tweaks the browserbuild script to avoid problems for
people developing on Microsoft Windows. Hat-tip to @philltew
for pointing this out, and submitting the fix.

Note that we'll need to make similar changes to the rollup
config in all patterns and plugins.