Fixed offset issue.
the offset property was not used appropriately. Instead of being offset from the point that is the start of the CoF, it was offset from the original starting point. This point has already been shifted by the margin property. 

This change also swaps the names of the cutonfoldFrom and cutonfoldTo points to be consistent with the to: and from: properties of the macro.