:chore: Made the shoulderSlopeReduction static
This option is used to create room for shoulder padding,
but that does not really make sense for a pattern like this.

The option is here because it's inherited from brian.
I assume @alfalyr started with the brian config as a base, and
didn't want to take out any options.

I think that's the smart play, except this option would probably
only confuse people in a context like this top.