:chore(titan): Overhaul of titan design
I worked with Deb on this to try and come up with a trouser block
that would be a good unisex base for all sort of patterns.

We started off with what Deb had been using to draft her own blocks
and ended up with something that worked for her, and only for her.

There were a lot of vague instructions in the draft, and a lot of
forks in the code where if this do that, if not do something
completely different.

The end result was a pattern that was very brittle, and also did
absolutely not work for a variety of sizes.

So, I basically just started over, threw all of that complicated
stuff out and this is not a more simple and streamlined block
that works for different models. It even passes the ant-man test :)

Now the question is, will it fit? That will be the next step.