fix(albert): Keep straps even when there's no SA
The width of the straps was calculated by multiplying with the `sa`
value.
As a result, when the user asks to draft a pattern without SA, the
straps drop off the pattern.

I've also change the way the width is calculated. It used to be based
on a percentage option that was then multiplied by SA.
While technically that follows the best practice of using percentages,
it won't scale with the model unless you multiply it with a measurement.