fix(components): Include snippets provided by theme plugin in Example component
Whereas most snippets are provided by build-time plugins that are
part of the plugin bundle, some of them are provided by the theme
plugin, which is a run-time plugin, and so we depend on the frontend
integration to make these available.

In the long(er) run, it might make more sense to move these to
another build-time plugin, but for the time being I've included
them into the Example component until we've clearly idendified
the ripple effects that moving them would have.

See #757 for follow-up of this.