feat(titan): Redrew the crotch and cross seam with new options
This commit adds the `crossSeamAngle` and `crotchSeamAngle` options
that control how the crotch- and cross seam reach the fork.

You can get the same result as what happened prior to this commit
by setting both options to zero.

However, their defaults are not zero, so this commit changes the
out-of-the-box pattern.