feat(paco): Extended heal ease option, added new options
This commit slightly increases the default heel ease value, but
also significantly increases the range of the option so that
people can make the leg wider.
This is in response to some people who found the original
design to tight on the calfs.

This also adds the crossSeamAngle and crotchSeamAngle
options from the upstream Titan block.

These angles allow you to influence how the crossSeam
and crotchSeam reach the fork.
Setting both of them to zero will give the same result
as the pattern before this commit.

Since their defaults are not zero, this will change the
out-of-the-box pattern (in addition to the added heal ease).