fix(cornelius): Side-step path splitting issue
There's an edge-case in Cornelius where things fall apart
when a path is split exactly on the endpoint of a curve
that is part of said path.

In a perfect world, core would handle this in a more robust way.
Until that time, this implements a workaround by checking for this
edge-case.

Closes #934