feat(brian): Allow shifting of the shoulder seam
This adds two new options to brian:

 - `s3Armhole`
 - `s3Collar`

s3 in this case is short for: Shoulder Seam Shift.

This option allows you to shift the shoulder seam forward
(by increasing the option) or backward (be decreasing it).
And this for the collar side and armhole/shoulder side.

This is in preparation of the feature request in #642.