breaking(simon): Replace yokeDart option by roundBack option
This option address fitting for a rounded back.
Something `yokeDart` used to do by adding a dart to the side of the
yoke.

However, in doing so `yokeDart` would change the lenght of the armhole
which is not ideal. The now `roundBack` option instead lengtens the back
in the middle, tapering off towards the sides. Thereby accomplishing
the same alteration without influencing the lenght of the armhole.