chore: Roll back changes to Holmes
This is an attempt to undo the changes made to the Holmes pattern
as the contributor of whom I merged the PR now doesn't want them
and I don't want to have code in the repo that people don't want to contribute.

Since I've updated translations, docs, changelog and whatnot since then,
this risks leaving things in a sub-optimal state but the damage is done.