fix(charlie): Turn waistbandWidht option into snapped pct
This in itself does not fix anything, but there's a complicated
regression because `@freesewing/snapseries` uses named exports.
And since Titan uses them, and is a Charlie dependency, things get
messy.

There's still some unresolved issues here, for example CRA doesn't
support ESM and thus the development environment for Charlie is still
effectively broken.

But by pulling snapseries in as a local dependency for charlie, at least
this should fix the problem on the website (once this is deployed)

Fixes #1460