fix(fs.dev): Force data-theme update with key prop
This is a replacement for the hack I implemented in
bb901c4d8a3b4b5e4f7569e574cb1c46c47a59a1

It accomplishes the same effect, but with just setting the key prop.
Sadly, it does not work with just setting the key prop to the theme
but forcing the key prop to be re-set by including the date does the
trick.

To be clear: This is still a hack and I'd love to avoid it (or just
understand the right way to handle this) but for now it will do.