Remove sleeve notch from Simon instructions.
The Simon instructions reference three sleeve notches: the double notch
on the back, the single notch on the front, and a non-existent single
notch at the top of the sleeve.

Example pattern missing notch: https://freesewing.org/create/simon/for/size-40-a/

Remove the non-existent notch from the the instrunctions and
corresponding images.