Added Sandy Instructions
This is a little messy, not going to lie. Skirts both fortunately and unfortunately have many ways to construct things. I have left certain things vauger than I would generally like, like Penelope but this is very much more a guide line rather than a set of rules. I had to split seamless and closure out from eachother as I feel it would be too confusing if they were combined. This feels like it is going to be an ongoing project than a this instructions are done but eh at least it is a start.