fix(i18n): Fix plugin translations
In a misguided attempt to keep the code DRY, the file that
builds the plugin translations was symlinked from various
languages to the English version.

Alas, due to the relative imports, this means that all
languages imported the English translations for plugins,
meaning that on-pattern translation was de-facto broken.

This fixes that.