Add deps from some Simon parts to back part.
I like to use the custom contents option but it doesn't work
for all parts in Simon. I believe this is due to the fact that the back
part sets values in the Store which other parts need. I first thought
about setting these values in shared but some of them are complex, for
example backArmholeToArmholePitch.