fix(tests): Pattern unit tests drafting of individual parts
After seeing #1768 I felt that we should have unit tests to make
sure that each pattern part drafts individually.

The I realized, *wait a minute, pretty sure I already added this*
and sure enough I had, but I did it wrong. Seemingly even I can't
remember that the *settings* are top-level properties in the config
passed to the constructure.

I've also removed the dependency on pattern-info (at least in this
part of the code) because I'd like to gradually ween us off this
package and deprecate it because it's just increasing the bundle
with info that's already available from the patterns anyway.

Also, it adds a required 2-step build since it needs all patterns
to be built before it can build itself. Since we're also gradually
getting rid of SASS it would be nice to have the ability to build
everything in one pass.