Honor Simon separate closure placket options
Simon has two options controlling each of the button plackets. One
option controls whether to use a separate placket, the other
controls which style of placket to use when using a built-in
placket. This change addresses an issue where the separate option was
ignored if the style option was set to seamless.