fix(docs) Holmes missing context
literally step above talks about the faced visor.
I also re-added the capitals. Ik it is not grammatically correct but the point is to make people double check what they are reading but also realise it is a sewing term. Now I am thinking I need to check the glossary and make sure some of the ones not showing in remark are on there. Mm.