feat: Avoid using bundler for loading YAML files
This adds a prebuild step to the i18n package that will
load the YAML files from disk and transform them into
Javascript files.

The files go in the `next` folder because for new I am keeping
this apart from the regular package.

The reason is that these files are optimized for NextJS and
not backwards compatible with the current way we handle translation.

In FreeSewing V3 we will use this and also switch to ESM-only
and named exports. For now, this is going to be kept apart but
used to load translations in our websites (currently handled by
a prebuild script).