feat(brian): Added waistToArmhole measurement
This is the start of an overhaul on how we construct the armhole in
Brian and patterns extending it.

Rather than base the armhole on the biceps circumference (and the
armholeDepthFactor option) we will use the waistToArmhole measurement
and the new armholeHeight option.

The armholeDepthFactor is hereby deprecated, although I've kept it in
the config for backwards compatibility.

Also note that the sleevecap now has a new sleevecapHeight option.
This option controls the overal height of the sleevecap independent
of the armhole. This is different from the previous armholeDepthFactor
option that would influence both armhole and sleevecap.

This change will have ripple effect on other patterns, those still need
to be made.