fix(lunetius-option-desc): to upper leg
changed the option description for `toUpperLeg` from "Below the hips" to "To the upper leg", as that is technically correct (it uses the `waist to upper leg` measurement). It sounds a bit more clunky, so opinions? (I'm also not sure if changing this here impacts anything else apart from the translation files, hence a PR and not a direct commit.)