Update back.js
This addresses the issue #2051. When the HPS to Bust becomes too small, the calculation of 'bustWidthBack' would throw an error because the line along which this was measured would no longer intersect with the side seam. This solution solves this by creating an artificial extension on the side seam that is used to create this point at which the measurement is made.
There is still an issue where if the value keeps decreasing, the side dart itself will no longer be on the side seam. This is an issue in front.js.