chore(pattern-info): Store prebuild info in repo
This is strictly speaking a bit silly to store build artifacts
in the repository. However, when building the freesewing.dev site
in CI we are depending on this data. And to generate it we need to
first install lerna and then build all packages so that the prebuild
step of this packages can happen.

So we can save ourselves a lot of hassle by simple shipping these
in the repo.