fix(markdown): Let comments be comments
While markdown does allow HTML, we do treat this content as MDX.
And the MDX parser tends to trip over HTML comments. Which is
typically not a big deal as people don't write them.

But then there's those who are too smart for their own good
like @tangerineshark :)

So I've replaced the HTML comments with the `Comment` component,
this leaves them visible, but I actually think that's better as
it shows it's a living document(ation).