chore: Supress webpack warnings about named imports from JSON
This suppresses a warning in webpack v5 when importing the
version from the package.json file like a named export.

This use to be fine in webpack v4 but in v5 it's deprecated to
be more inline with the way it works in NodeJS.

That being said, it's very usefull as there's no need to pull in
the entire package.json file.
Futhermore, esbuild supports this out of the box so we also use
it in our build scripts.

So for now I'm supressing this warning. In FreeSewing v3 we'll
re-evaluate this.