feat(config-helpers): Renamed snapseries to config-helpers
It would be useful to have a place where we can place methods
like the pctBasedOn method that we now import from core, even though
it has nothing to do with core.

So I've decided to rename this packet since it's essentially also
things that facilitate pattern configuration (files).

Since this is a very new package, and I am 100% certain that we
(FreeSewing itself) are the only ones using it, no need to major
bump this. The most recent snapseries package will remain available
for people who want to use it.

I'll update all patterns accordingly.