chore(models): Keep neckstimate in bundle coz Gatsby
This legacy freesewing.org fails to build with the latest (pre-)release
because it can't seem to resolve the neckstimate function from the
models package.

I've setup models to include the neckstimate method in the bundle as
that addresses the issue, and it's hardly a big deal.

Going forward with the new site, we'll drop the entire utils package
anyway.