fix(shared): Don't force view to measurements by default
This is a bit of an annoyance since it forces you to the measurment
view, even for designs that have no required measurments.

So I've switched it to the draft view, and it should be the useEffect
code that detects whether required measurments are lacking and then
update the view.

FYI: @eriese