fix(shared): Avoid flash of draft error before re-render
When measurments are missing, the useEffect hook will rerender to show
the measurements page. However, before that happens there's a brief
flash of the (failed) draft because it errors without the measurements.

This prevents that by passing down whether or not all required
measurements are available, and if not the draft returns null