fix(shared): Don't import from core source files
Importing from source files (of published packages) works fine
in the monorepo, but breaks outside of it.
Specifically, it breaks in the standalone development environment
which spins up the lab outside the monorepo.

For a situation like this, you need to make sure we have a
named export that's available from the built package.