fix: Broken link in README's
As reported in #2605 there's a broken link in the README's that points
to an old overview of all published packages on freesewing.dev.

Since we migrated to NextJS that page no longer exists, futhermore we've
moved away from the single `packages` workspace and now have one for
`packages`, `designs`, `plugins`, and `sites`. So I've just dropped the
link and rephrased the sentence.