fix(simone): Don't mutate the simon config
The changes that were added recently for the `bustAlignedButtons` option
were causing problems when both simon and simone are loaded.

That's because the simone config adds the option to (an object of) the
simon config. Because the reference to this object is the same in both
configs, that means the option also appears on Simon, which causes
issues.

I've (for now) merely duplicated this part of the config to side-step
the issue. For now, because we'll handle this differently/better in v3.