chore: Removed examples as it's no longer needed
We no longer need a dedicated examples design.
It was used to showcase examples in the dev docs.
But the new dev site Example component will use the actual example code
to generate the example. Which is much better since we don't have to
keep the example code and code of the the example design in sync.