wip(core); Idea to facilitate checking for store methods.
Calling `store.some.method()` is risky because what if the method is not
available (as it's provided by a plugin).

The correct way to handle that is to check whether it's a method:

```js
if (typeof store.some.method === 'function') {
  store.some.method()
}
```

But that will get tired pretty soon.

This proposal implements the following alternative check:

```js
if (store.hasMethod('some.method')) store.some.method()
```

It's just a quick idea, this needs tests and so on.