chore(org): Create the use-account hook
I am refactoring hooks, breaking up the *god hook* that is useApp.

I am also switching to `use-persisted-storage` as a dependency
over our own useStorage implementation as that has been the source
of some bugs lately and this one seems to do the job just fine
so that's one less thing to worry about.