wip(core): Work on supported part-level transforms in layouting
This is some initial work to support part-level (SVG) transforms when
layouting the pattern.

It updates the method that calculates the bounding box to calculate a
bounding box with transforms applied.
It also adds two new methods to utils to help with that, methods that
could be useful for other people trying to do fancy transform stuff.