feat(shared): Add Mermaid support to MDX pipeline
At the request of @marpants this adds support for mermaid diagrams to
our shared MDX pipeline. Diagrams can be created by using a fenced code
block using the `mermaid` language.

I did not verify to what extend this bloats the bundle. Perhaps that's
something we need to look into later and consider the trade-off. For
now, my main concern is to keep a new contributor happy, so here it goes
:)