breaking(i18n): The @freesewing/i18n package is no more
This concludes the initial migration of our translation to the new v3
structure.

We no longer publish a `@freesewing/i18n` package with the translations
for our project, instead each design has a `i18n` names export that
holds the strings for that design.

The plugins still need to be ported, and `plugin-i18n` needs to be
adapted, but the main work is done.