chore: Privacy Notice update
This privacy notice update was triggered by a need to remove the
ambigiuty about what 'last login' means.

While at it, I have restructured it so it is a single page rather than
spread out over various pages, and restructured it a bit as well as
added the warrent canaries.

The underlying policy is not changed, it is more a matter of bringing
the language in line with the upcoming v3 backend and the way we store
data.