chore: Further privacy notice updates
Apart from fixing broken links to other documentation, this consolidates
the previous 'profile' and 'people' consent into a unified 'account'
consent.

The reason is historic, we used to require an account to generate any
sort of pattern. This was the 'profile' consent. Then, in addition to
that you could add your own measurements and generate patterns for
those, this was the 'people' consent.

However, we have changed the account requirement in the sense that you
can now generate any pattern you want without an account. You just
cannot use your own personal measurements.

As such, the distinction is no longer useful as an account is completely
useless (as in: gives you nothing extra in comparison to not having an
account) without consent to process those personal measurements.

So, one consent type now.