fix(plugin-annotations): Set bias/cutonfold explicity in cutlist
The cutlist methods were being clever by deducing whether a part should
be cut on the fold from info set by the cutonfold macro.

This is a footgun, because this only works when the cutlist is set after
the cutonfold macro is run.

We can't expect designers to know they should add this annotation before
that annotation so now you need to add it explicitly.

The current keys to pass are:

 - cut
 - from (used to be material)
 - onBias (used to be bias)
 - onFold (replaced ignoreOnFold)