chore: Change mustache delimiter to keep tempaltes as .mjs
The problem with auto-generated pages is that prettier will change
the formatting ever so slightly in the commit hook.
Which means that next time they are generated, git will see a bunch
of meaningless changes.

So, in changing the mustache delimter to $$ (valid JS) we can keep the
template files as .mjs and let prettier format them too, which I hope
will mean that we can sidestep this constant churn.