chore(markdown): Remove sample images from docs
Since the website now supports this, there's no need to keep this in the
documentation. Especially since it's not great documentation as it's not
that easy to read the changes (depending on the option).

In addition, maintaining all these SVG files is not a walk in the part,
plus they are not theme-compatible.

Options should be properly documented, and those who want to see how the
option affacts a pattern can check that in the pattern editor.