chore(backend): Remove (c)set from patterns, img from users
This removes some fields from the database and the logic for handling
them. Specifically:

- Do not keep a set or cset reference for patterns. Doing so will only
  raise expectations for users as a link between (c)set and pattern
  might signal that changes to the (c)set would somehow also update the
  pattern. Once a pattern is saved, it is what it is, so no point
  keeping a link to the (c)set. We may want to look at how to add a note
  about which (c)set the pattern was generated for, but I don't think
  that's crucial.
- Do not keep an img field in the user account, cset, or open pack,
  since they have a deterministic image ID anyway.

This also adds and info field to cset and option pack which is an
internal field allowing us to keep (non-public) notes about where the
record comes from. This will be important when we add a new mesurement
and we have to contqct all these people for that extra measurement (for
example).