fix: Allow clicks in modal wrapper content
This raises the content inside the modal wrapper and stops propagation
of clicks. This fixes #5527 which closed the modal when people were
trying to select a different tab in the modal content.

On desktop, you can still close the modal by clicking outside of the
content. However, on mobile, the modal fills the entire screen so there
is no way to click outside and close the modal.
So this also includes a change that adds a close button to the bottom
right corner (only shown on mobile).