feat(backend): Add getSubscribers route
This adds routes to get the newsletter subscriber list.
This is used by the script that sends out the newsletters.
Minimum RBAC level required is support (8). This is because you cannot
generate an admin-level (9) API key via the website.

But it's probably something that's ok for people in support roles to be
able to do.